Change field `crc` in characters.addons to int(32) to support proper values.
--HG--
branch : trunk

+++ sql/characters.sql
@@ -49,7 +49,7 @@ DROP TABLE IF EXISTS `addons`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `addons` (
   `name` varchar(120) NOT NULL default '',
-  `crc` mediumint(32) unsigned NOT NULL default '0',
+  `crc` int(32) unsigned NOT NULL default '0',
   PRIMARY KEY (`name`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 ROW_FORMAT=DYNAMIC COMMENT='Addons';
 /*!40101 SET character_set_client = @saved_cs_client */;
